--- snakecord/clients/__init__.py
@@ -1,1 +1,0 @@
-from .user import *  # noqa: F401, F403
