--- snakecord/clients/user/__init__.py
@@ -1,2 +1,0 @@
-from .client import UserClient  # noqa: F401
-from .manager import UserClientEvents, UserClientManager  # noqa: F401
