+++ snakecord/utils/undefined.py
@@ -2,11 +2,11 @@ __all__ = ('undefined',)
 
 
 class _Undefined:
-    def __repr__(self):
-        return '<undefined>'
-
     def __bool__(self):
         return False
+
+    def __repr__(self):
+        return '<undefined>'
 
 
 undefined = _Undefined()
